Move `FeedImagesMapper` to its own file
